Create cpu_gpu_opencl.cpp
We used OpenCL to calculate NDVI by leveraging both the CPU and GPU of my system. Our CPU is an Intel Core i5-5300U with 2 cores, 4 threads, and a max clock speed of 2.9 GHz, while the GPU is an Intel HD Graphics 5500. OpenCL allowed us to parallelize the computation, with the CPU handling smaller tasks and the GPU processing image data in parallel for faster performance. We were able to efficiently calculate the NDVI from multispectral images, improving speed comparing to CPU only approach.